requirements.txt
python-telegram-bot==20.6
sqlalchemy
python-dotenv
openai
requests
pytz
alembic
apscheduler
reportlab
arabic-reshaper
python-bidi
